+++ examples/ambiguous1/goutil/bsr/bsr.go
@@ -27,8 +27,6 @@ package bsr
 
 import (
 	"fmt"
-	"os"
-	"sort"
 	"strings"
 	"unicode/utf8"
 
@@ -42,16 +40,23 @@ type bsr interface {
 }
 
 var (
-	set *bsrSet
+	set      *bsrSet
 	startSym string
 )
 
 type bsrSet struct {
 	slotEntries   map[BSR]bool
-	ignoredSlots   map[BSR]bool
+	ntSlotEntries map[ntSlot][]BSR
+	ignoredSlots  map[BSR]bool
 	stringEntries map[stringBSR]bool
 	rightExtent   int
-	I			  []byte
+	I             []byte
+}
+
+type ntSlot struct {
+	nt          string
+	leftExtent  int
+	rightExtent int
 }
 
 // BSR is the binary subtree representation of a parsed nonterminal
@@ -72,10 +77,11 @@ type stringBSR struct {
 func newSet(I []byte) *bsrSet {
 	return &bsrSet{
 		slotEntries:   make(map[BSR]bool),
-		ignoredSlots:   make(map[BSR]bool),
+		ntSlotEntries: make(map[ntSlot][]BSR),
+		ignoredSlots:  make(map[BSR]bool),
 		stringEntries: make(map[stringBSR]bool),
-		rightExtent:	0,
-		I: I,
+		rightExtent:   0,
+		I:             I,
 	}
 }
 
@@ -94,7 +100,7 @@ func Add(l slot.Label, i, k, j int) {
 }
 
 func AddEmpty(l slot.Label, i int) {
-	insert(stringBSR{l, i, i, i})
+	insert(BSR{l, i, i, i})
 }
 
 func Contain(nt string, left, right int) bool {
@@ -123,7 +129,7 @@ func GetRightExtent() int {
 	return set.rightExtent
 }
 
-// GetRoot returns the root of the parse tree of an unambiguous parse. 
+// GetRoot returns the root of the parse tree of an unambiguous parse.
 // GetRoot fails if the parse was ambiguous. Use GetRoots() for ambiguous parses.
 func GetRoot() BSR {
 	rts := GetRoots()
@@ -138,16 +144,6 @@ func GetRoots() (roots []BSR) {
 	for s, _ := range set.slotEntries {
 		if s.Label.Head() == startSym && s.leftExtent == 0 && s.rightExtent == set.rightExtent {
 			roots = append(roots, s)
-		}
-	}
-	return
-}
-
-// GetNTSlot returns all slot entries of the NT, 'nt'
-func GetNTSlot(nt string) (slots []BSR) {
-	for bsr := range set.slotEntries {
-		if bsr.Label.Head() == nt {
-			slots = append(slots, bsr)
 		}
 	}
 	return
@@ -176,6 +172,8 @@ func insert(bsr bsr) {
 	switch s := bsr.(type) {
 	case BSR:
 		set.slotEntries[s] = true
+		nt := ntSlot{s.Label.Head(), s.leftExtent, s.rightExtent}
+		set.ntSlotEntries[nt] = append(set.ntSlotEntries[nt], s)
 	case stringBSR:
 		set.stringEntries[s] = true
 	default:
@@ -202,8 +200,8 @@ func (b BSR) GetNTChild(nt string, i int) BSR {
 	return bsrs[0]
 }
 
-// GetNTChildI returns the BSR of NT symbol[i] in s.
-// GetNTChildI fails if s has ambiguous subtrees of NT i.
+// GetNTChildI returns the BSR of NT symbol[i] in the BSR set.
+// GetNTChildI fails if the BSR set has ambiguous subtrees of NT i.
 func (b BSR) GetNTChildI(i int) BSR {
 	bsrs := b.GetNTChildrenI(i)
 	if len(bsrs) != 1 {
@@ -259,10 +257,35 @@ func (b BSR) GetString() string {
 	return string(set.I[b.LeftExtent():b.RightExtent()])
 }
 
+// Ignore removes NT slot 'b' from the BSR set. Ignore() is typically called by 
+// disambiguration code to remove an ambiguous BSR entry.
 func (b BSR) Ignore() {
 	// fmt.Printf("bsr.Ignore %s\n", b)
 	delete(set.slotEntries, b)
+	deleteNTSlotEntry(b)
 	set.ignoredSlots[b] = true
+}
+
+func deleteNTSlotEntry(b BSR) {
+	// fmt.Printf("deletNTSlotEntry(%s)\n", b)
+	nts := ntSlot{b.Label.Head(), b.leftExtent, b.rightExtent}
+	slots := set.ntSlotEntries[nts]
+	bi := -1
+	for i, s := range slots {
+		// fmt.Println("", i, ":", s)
+		if s == b {
+			if bi != -1 {
+				panic("Duplicate")
+			}
+			bi = i
+			// fmt.Println("  bi", i)
+		}
+	}
+	slots1 := slots[0:bi]
+	slots1 = append(slots1, slots[bi+1:]...)
+	// fmt.Println("  slots", slots)
+	// fmt.Println("  slots1", slots1)
+	set.ntSlotEntries[nts] = slots1
 }
 
 func (s BSR) LeftExtent() int {
@@ -297,6 +320,7 @@ func (s stringBSR) Empty() bool {
 	return s.leftExtent == s.pivot && s.pivot == s.rightExtent
 }
 
+// String returns a string representation of s
 func (s stringBSR) String() string {
 	// fmt.Printf("bsr.stringBSR.stringBSR(): %s, %d, %d, %d\n",
 	// 	s.Label.Symbols(), s.leftExtent, s.pivot, s.rightExtent)
@@ -306,74 +330,18 @@ func (s stringBSR) String() string {
 		s.rightExtent, set.I[s.LeftExtent():s.RightExtent()])
 }
 
-func Dump() {
-	DumpSlots()
-	DumpStrings()
-}
-
-func DumpSlots() {
-	fmt.Printf("Slots (%d)\n", len(getSlots()))
-	for _, s := range getSlots() {
-		DumpSlot(s)
-	}
-}
-
-func DumpSlot(s BSR) {
-	fmt.Println(s)
-}
-
-func DumpStrings() {
-	fmt.Printf("Strings(%d)\n", len(getStrings()))
-	for _, s := range getStrings() {
-		dumpString(s)
-	}
-}
-
-func dumpString(s stringBSR) {
-	fmt.Println(s)
-}
-
-func getSlots() (slots []BSR) {
-	for s := range set.slotEntries {
-		slots = append(slots, s)
-	}
-	sort.Slice(slots,
-		func(i, j int) bool {
-			return slots[i].Label < slots[j].Label
-		})
-	return
-}
-
-func getStrings() (strings []stringBSR) {
-	for s := range set.stringEntries {
-		strings = append(strings, s)
-	}
-	sort.Slice(strings,
-		func(i, j int) bool {
-			return strings[i].Label < strings[j].Label
-		})
-	return
-}
-
 func getNTSlot(nt string, leftExtent, rightExtent int) (bsrs []BSR) {
-	for sl, _ := range set.slotEntries {
-		if sl.Label.Head() == nt && sl.leftExtent == leftExtent && sl.rightExtent == rightExtent {
-			bsrs = append(bsrs, sl)
-		}
-	}
-	return
+	return set.ntSlotEntries[ntSlot{nt, leftExtent, rightExtent}]
 }
 
 func fail(b BSR, format string, a ...interface{}) {
 	msg := fmt.Sprintf(format, a...)
 	line, col := getLineColumn(b.LeftExtent(), set.I)
-	fmt.Printf("Error in BSR: %s at line %d col %d\n", msg, line, col)
-	os.Exit(1)
+	panic(fmt.Sprintf("Error in BSR: %s at line %d col %d\n", msg, line, col))
 }
 
 func failf(format string, args ...interface{}) {
-	fmt.Printf("Error in BSR: %s\n", fmt.Sprintf(format, args...))
-	os.Exit(1)
+	panic(fmt.Sprintf("Error in BSR: %s\n", fmt.Sprintf(format, args...)))
 }
 
 func decodeRune(str []byte) (string, rune, int) {
@@ -410,4 +378,3 @@ func getLineColumn(cI int, I []byte) (line, col int) {
 	}
 	return
 }
-