+++ examples/ambiguous1/goutil/stringset/stringset.go
@@ -21,6 +21,7 @@ import (
 	"bytes"
 	"fmt"
 	"sort"
+	"strings"
 )
 
 type StringSet struct {